+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.3.3
+Improving security. 
+Some Ips are banned now if submission is not valid and cannot resubmit.
+Ips are banned if they have same email or same message as another banned ip.
+Cleaning all plugins data on deactivation.
+Some refactoring.
+Updating some translation.
+
 ## 1.3.2
 Changing wp_head hook call.